feat: Refactorisation du code avec mise à jour des chemins d'importation, suppression de fichiers obsolètes et ajout d'un nouveau script de diagnostic
