+++ src/envoy_api.py
@@ -10,7 +10,7 @@ from typing import Any, Dict, Optional
 
 import aiohttp
 
-from const import (
+from src.const import (
     ENLIGHTEN_LOGIN_URL,
     ENTREZ_TOKEN_URL,
     ENVOY_AUTH_CHECK_ENDPOINT,
